pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ceda-services-portal"
version = "0.2.0"
authors = [{name = "William Tucker", email = "william.tucker@stfc.ac.uk"}]
maintainers = [{name = "William Tucker", email = "william.tucker@stfc.ac.uk"}]
license = {text = "BSD - See ceda_example/LICENSE file for details"}
description = "Django application for registering and managing CEDA users."
classifiers = [
    "Development Status :: 2 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: BSD License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
]
# Run this:
# pip-compile --allow-unsafe --no-annotate --output-file=requirements.txt --resolver=backtracking --upgrade pyproject.toml
dependencies = [
    "Django>=3.2.19,<3.3",
    "crypto-cookie @ git+https://github.com/cedadev/crypto-cookie.git@b92dd9e6cc910d48d9380222c9273c8929666980",
    "django-cookie-law==2.2.0",
    "django-countries>=7.2.1,<7.6.0",
    "django-crispy-forms>=1.14.0,<2.0.0",
    "django-markdown-deux>=1.0.5,<2.0.0",
    "django-oidc-extras @ git+https://github.com/cedadev/django-oidc-extras.git@7865641af8e0886bca244e2fc9ab27221f518ef8",
    "django-polymorphic>=3.1.0,<4.0.0",
    "django-widget-tweaks>=1.4.12,<2.0.0",
    "djangorestframework>=3.13.1,<4.0.0",
    "fwtheme-django @ git+https://github.com/cedadev/fwtheme-django.git@bootstrap5",
    "jasmin-django-utils @ git+https://github.com/cedadev/jasmin-django-utils.git@v1.0.9",
    "jasmin-ldap @ git+https://github.com/cedadev/jasmin-ldap.git@v1.0.0",
    "jasmin-ldap-django @ git+https://github.com/cedadev/jasmin-ldap-django.git@v1.0.1",
    "jasmin-metadata @ git+https://github.com/cedadev/jasmin-metadata.git@v1.0.2",
    "jasmin-notifications @ git+https://github.com/cedadev/jasmin-notifications.git@v1.2.7",
    "jasmin-services[keycloak] @ git+https://github.com/cedadev/jasmin-services.git@321361fb58de3ed2cf6be10377d74d8b1ebdda14",
    "mozilla-django-oidc>=2.0.0,<3.0.0",
    "oauthlib>=3.2.0,<4.0.0",
    "pika>=1.2.0,<2.0.0",
    "psycopg2-binary>=2.9.3,<3.0.0",
    "python-dateutil>=2.8.2,<3.0.0",
    "requests-oauthlib>=1.3.1,<2.0.0",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
Homepage = "https://github.com/cedadev/ceda-services-portal"

[tool.black]
target-version = ['py310', 'py311']

[tool.isort]
profile = "black"

[tool.pylint.main]
load-plugins = 'pylint_django'

[tool.setuptools]
zip-safe = false
include-package-data = false

[tool.setuptools.packages]
find = {namespaces = false}

[tool.setuptools.package-data]
services_portal_site = ["templates/**", "static/**"]
account = ["templates/**", "static/**"]
